refactor(livekit): extract shared /rtc proxy helpers
_web_server.py and _token_server.py duplicated ~50 lines of
LiveKit-facing logic (the /token JWT issuer, /rtc/validate HTTP
proxy, and /rtc WebSocket pump). Both servers expose the same
proxy surface to browser clients with slightly different HTTP
shells; only the shells differ.

Extract:
- _lk_proxy.proxy_validate(client, base, request) — /rtc/validate
- _lk_proxy.pump_rtc_ws(client_ws, lk_internal_ws) — /rtc WS pump
- _token.make_client_token(cfg, identity, ttl=None) — accepts None
  to skip the explicit TTL, used by both web servers (preserving
  prior LiveKit-default-TTL behavior).

The defensive `msg_type != "websocket.receive"` guard from
_web_server.py is now applied uniformly via _lk_proxy. Public API
(LiveKitConnector / make_client_token) and existing TTL behavior
for non-web callers (24h default) are unchanged.

Closes the SonarQube duplication finding on _web_server.py (47.9%).
