+++ server-runtime/xr_media_hub/transport/livekit/_token_server.py
@@ -19,12 +19,12 @@ import logging
 
 import httpx
 import uvicorn
-import websockets
 from fastapi import FastAPI, Query, Request, WebSocket
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.responses import Response
-from livekit.api import AccessToken, VideoGrants
 
+from . import _lk_proxy
+from ._token import make_client_token
 from .config import LiveKitConnectorConfig
 
 log = logging.getLogger(__name__)
@@ -44,67 +44,17 @@ def build_app(cfg: LiveKitConnectorConfig) -> FastAPI:
 
     @app.get("/token")
     async def get_token(identity: str = Query(default="browser-user")) -> dict:
-        token = (
-            AccessToken(cfg.api_key, cfg.api_secret)
-            .with_identity(identity)
-            .with_name(identity)
-            .with_grants(VideoGrants(room_join=True, room=cfg.room_name))
-            .to_jwt()
-        )
+        token = make_client_token(cfg, identity=identity, ttl=None)
         return {"token": token, "room": cfg.room_name, "url": cfg.token_server_url}
 
     @app.get("/rtc/validate")
     async def rtc_validate(request: Request) -> Response:
-        qs = str(request.url.query)
         async with httpx.AsyncClient() as client:
-            r = await client.get(f"{lk_internal_http}/rtc/validate?{qs}")
-        return Response(
-            content=r.content,
-            status_code=r.status_code,
-            media_type=r.headers.get("content-type", "text/plain"),
-        )
+            return await _lk_proxy.proxy_validate(client, lk_internal_http, request)
 
     @app.websocket("/rtc")
     async def rtc_ws_proxy(client_ws: WebSocket) -> None:
-        qs = client_ws.scope.get("query_string", b"").decode()
-        target = f"{lk_internal_ws}/rtc?{qs}"
-        await client_ws.accept()
-        try:
-            async with websockets.connect(target, max_size=None) as lk_ws:
-
-                async def c2l() -> None:
-                    try:
-                        while True:
-                            msg = await client_ws.receive()
-                            if msg["type"] == "websocket.disconnect":
-                                break
-                            if msg.get("bytes"):
-                                await lk_ws.send(msg["bytes"])
-                            elif msg.get("text"):
-                                await lk_ws.send(msg["text"])
-                    except Exception:
-                        pass
-                    finally:
-                        await lk_ws.close()
-
-                async def l2c() -> None:
-                    try:
-                        async for frame in lk_ws:
-                            if isinstance(frame, bytes):
-                                await client_ws.send_bytes(frame)
-                            else:
-                                await client_ws.send_text(frame)
-                    except Exception:
-                        pass
-
-                await asyncio.gather(c2l(), l2c(), return_exceptions=True)
-        except Exception as exc:
-            log.debug("WS proxy /rtc error: %s", exc)
-        finally:
-            try:
-                await client_ws.close()
-            except Exception:
-                pass
+        await _lk_proxy.pump_rtc_ws(client_ws, lk_internal_ws)
 
     if cfg.browser_dir:
         from fastapi.staticfiles import StaticFiles
